perf: entry
