+++ projects/app/src/pages/_app.tsx
@@ -1,103 +1,17 @@
-import { useEffect, useState } from 'react';
 import type { AppProps } from 'next/app';
 import Script from 'next/script';
 import Head from 'next/head';
-import { ChakraProvider, ColorModeScript } from '@chakra-ui/react';
 import Layout from '@/components/Layout';
-import { theme } from '@fastgpt/web/styles/theme';
-import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
-import NProgress from 'nprogress'; //nprogress module
-import Router from 'next/router';
-import { clientInitData } from '@/web/common/system/staticData';
-import { appWithTranslation, useTranslation } from 'next-i18next';
-import { useRouter } from 'next/router';
-import { useSystemStore } from '@/web/common/system/useSystemStore';
-import type { FastGPTFeConfigsType } from '@fastgpt/global/common/system/types/index.d';
-import { change2DefaultLng, setLngStore } from '@/web/common/utils/i18n';
+import { appWithTranslation } from 'next-i18next';
 
-import 'nprogress/nprogress.css';
+import QueryClientContext from '@/web/context/QueryClient';
+import ChakraUIContext from '@/web/context/ChakraUI';
+import { useInitApp } from '@/web/context/useInitApp';
+
 import '@/web/styles/reset.scss';
 
-//Binding events.
-Router.events.on('routeChangeStart', () => NProgress.start());
-Router.events.on('routeChangeComplete', () => NProgress.done());
-Router.events.on('routeChangeError', () => NProgress.done());
-
-// Create a client
-const queryClient = new QueryClient({
-  defaultOptions: {
-    queries: {
-      keepPreviousData: true,
-      refetchOnWindowFocus: false,
-      retry: false,
-      cacheTime: 10
-    }
-  }
-});
-
 function App({ Component, pageProps }: AppProps) {
-  const router = useRouter();
-  const { hiId } = router.query as { hiId?: string };
-  const { i18n } = useTranslation();
-  const { loadGitStar, setInitd, feConfigs } = useSystemStore();
-  const [scripts, setScripts] = useState<FastGPTFeConfigsType['scripts']>([]);
-  const [title, setTitle] = useState(process.env.SYSTEM_NAME || 'AI');
-
-  useEffect(() => {
-    // get init data
-    (async () => {
-      const {
-        feConfigs: { scripts, isPlus, show_git, systemTitle }
-      } = await clientInitData();
-
-      setTitle(systemTitle || 'FastGPT');
-
-      // log fastgpt
-      if (!isPlus) {
-        console.log(
-          '%cWelcome to FastGPT',
-          'font-family:Arial; color:#3370ff ; font-size:18px; font-weight:bold;',
-          `GitHub：https://github.com/labring/FastGPT`
-        );
-      }
-      if (show_git) {
-        loadGitStar();
-      }
-
-      setScripts(scripts || []);
-      setInitd();
-    })();
-
-    // add window error track
-    window.onerror = function (msg, url) {
-      window.umami?.track('windowError', {
-        device: {
-          userAgent: navigator.userAgent,
-          platform: navigator.platform,
-          appName: navigator.appName
-        },
-        msg,
-        url
-      });
-    };
-
-    return () => {
-      window.onerror = null;
-    };
-  }, []);
-
-  useEffect(() => {
-    // get default language
-    const targetLng = change2DefaultLng(i18n.language);
-    if (targetLng) {
-      setLngStore(targetLng);
-      router.replace(router.asPath, undefined, { locale: targetLng });
-    }
-  }, []);
-
-  useEffect(() => {
-    hiId && localStorage.setItem('inviterId', hiId);
-  }, [hiId]);
+  const { feConfigs, scripts, title } = useInitApp();
 
   return (
     <>
@@ -115,17 +29,15 @@ function App({ Component, pageProps }: AppProps) {
       </Head>
       {scripts?.map((item, i) => <Script key={i} strategy="lazyOnload" {...item}></Script>)}
 
-      <QueryClientProvider client={queryClient}>
-        <ChakraProvider theme={theme}>
-          <ColorModeScript initialColorMode={theme.config.initialColorMode} />
+      <QueryClientContext>
+        <ChakraUIContext>
           <Layout>
             <Component {...pageProps} />
           </Layout>
-        </ChakraProvider>
-      </QueryClientProvider>
+        </ChakraUIContext>
+      </QueryClientContext>
     </>
   );
 }
 
-// @ts-ignore
 export default appWithTranslation(App);
